pull content from newsandinfohomepagesection into its parent component, and update strapi, fix variants
