+++ next/src/components/page-contents/HomepageContent.tsx
@@ -1,7 +1,12 @@
+import HorizontalDivider from '@/src/components/common/Divider/HorizontalDivider'
 import Waves from '@/src/components/common/Waves/Waves'
+import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { useHomepageContext } from '@/src/components/providers/HomepageContextProvider'
+import AnnouncementBannerSection from '@/src/components/sections/homepage-sections/AnnouncementBannerSection'
+import HighlightsHomepageSection from '@/src/components/sections/homepage-sections/HighlightsHomepageSection'
+import HomepageTabs from '@/src/components/sections/homepage-sections/HomepageTabs/HomepageTabs'
 import InbaHomepageSection from '@/src/components/sections/homepage-sections/InbaHomepageSection'
-import NewsAndInfoHomepageSection from '@/src/components/sections/homepage-sections/NewsAndInfoHomepageSection'
+import MayorAndCouncilHomepageSection from '@/src/components/sections/homepage-sections/MayorAndCouncilHomepageSection'
 import TopServicesHomepageSection from '@/src/components/sections/homepage-sections/TopServicesHomepageSection'
 import WelcomeHomepageSection from '@/src/components/sections/homepage-sections/WelcomeHomepageSection'
 import TootootEventsSection from '@/src/components/sections/TootootEventsSection'
@@ -17,11 +22,22 @@ const HomepageContent = () => {
 
       <Waves waveColor="var(--color-grey-50)" wavePosition="top" />
 
-      <NewsAndInfoHomepageSection
-        announcementBannerSectionProps={
-          announcementBannerSection ? announcementBannerSection : null
-        }
-      />
+      <SectionContainer className="bg-background-passive-primary pb-14">
+        <HighlightsHomepageSection />
+
+        <HomepageTabs className="lg:mt-10" />
+
+        {announcementBannerSection ? (
+          <AnnouncementBannerSection
+            {...announcementBannerSection}
+            className="mt-8 mb-7.5 lg:mt-18 lg:mb-8.5"
+          />
+        ) : (
+          <HorizontalDivider className="mt-8 lg:mt-18" />
+        )}
+
+        <MayorAndCouncilHomepageSection className="mt-8 lg:mt-18" />
+      </SectionContainer>
 
       <Waves wavePosition="bottom" waveColor="var(--color-grey-50)" />
