+++ strapi/config/sync/core-store.plugin_content_manager_configuration_components##sections.announcement-banner.json
@@ -54,8 +54,7 @@
           "description": "",
           "placeholder": "",
           "visible": true,
-          "editable": true,
-          "mainField": "label"
+          "editable": true
         },
         "list": {
           "label": "link",
@@ -87,6 +86,12 @@
       }
     },
     "layouts": {
+      "list": [
+        "id",
+        "title",
+        "text",
+        "link"
+      ],
       "edit": [
         [
           {
@@ -110,12 +115,6 @@
             "size": 12
           }
         ]
-      ],
-      "list": [
-        "id",
-        "title",
-        "text",
-        "link"
       ]
     },
     "uid": "sections.announcement-banner",
