+++ strapi/src/components/sections/announcement-banner.json
@@ -20,6 +20,7 @@
     },
     "variant": {
       "type": "enumeration",
+      "required": true,
       "default": "dark",
       "enum": [
         "dark",
